Refact Long Method and bad structure
