+++ Checkers/Services/GameSaves/GameSaver.cs
@@ -92,55 +92,4 @@ namespace Checkers.Services
             }
         }
     }
-
-    public class GameState
-    {
-        public Board Board { get; set; }
-        public PlayerType CurrentPlayer { get; set; }
-        public PieceInfo SelectedPiece { get; set; }
-        public List<MoveInfo> ValidMoves { get; set; }
-        public bool IsGameOver { get; set; }
-        public PlayerType Winner { get; set; }
-    }
-
-    public class PieceInfo
-    {
-        public int Row { get; set; }
-        public int Col { get; set; }
-        public PlayerType Player { get; set; }
-        public PieceType Type { get; set; }
-
-        public PieceInfo() { }
-        public PieceInfo(Piece piece)
-        {
-            Row = piece.Row;
-            Col = piece.Col;
-            Player = piece.Player;
-            Type = piece.Type;
-        }
-    }
-
-    public class MoveInfo
-    {
-        public int PieceRow { get; set; }
-        public int PieceCol { get; set; }
-        public int ToRow { get; set; }
-        public int ToCol { get; set; }
-        public int CapturedPieceRow { get; set; } = -1;
-        public int CapturedPieceCol { get; set; } = -1;
-
-        public MoveInfo() { }
-        public MoveInfo(Move move)
-        {
-            PieceRow = move.Piece.Row;
-            PieceCol = move.Piece.Col;
-            ToRow = move.ToRow;
-            ToCol = move.ToCol;
-            if (move.CapturedPiece != null)
-            {
-                CapturedPieceRow = move.CapturedPiece.Row;
-                CapturedPieceCol = move.CapturedPiece.Col;
-            }
-        }
-    }
 }
